refactor out nested scaffold
