+++ app/src/main/java/com/app/whakaara/ui/navigation/NavGraph.kt
@@ -6,6 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.navigation
+import com.app.whakaara.state.events.AppViewModels
 import com.whakaara.core.LeafScreen
 import com.whakaara.core.RootScreen
 import com.whakaara.feature.alarm.AlarmViewModel
@@ -19,9 +20,7 @@ import com.whakaara.onboarding.navigation.onboardingScreen
 @Composable
 fun NavGraph(
     navController: NavHostController,
-    stopwatchViewModel: StopwatchViewModel,
-    timerViewModel: TimerViewModel,
-    alarmViewModel: AlarmViewModel,
+    viewModels: AppViewModels,
     shouldShowOnboarding: Boolean = false
 ) {
     NavHost(
@@ -34,13 +33,13 @@ fun NavGraph(
     ) {
         addOnboardingRoute(navController)
         addAlarmRoute(
-            viewModel = alarmViewModel
+            viewModel = viewModels.alarm
         )
         addTimerRoute(
-            viewModel = timerViewModel
+            viewModel = viewModels.timer
         )
         addStopwatchRoute(
-            viewModel = stopwatchViewModel
+            viewModel = viewModels.stopwatch
         )
     }
 }
