+++ feature/alarm/src/androidTest/kotlin/com/whakaara/feature/alarm/AlarmScreenTest.kt
@@ -45,7 +45,6 @@ class AlarmScreenTest {
                 AlarmScreen(
                     alarms = listOf(firstAlarm, secondAlarm),
                     preferencesState = PreferencesState(),
-                    create = {},
                     delete = {},
                     disable = {},
                     enable = {},
