+++ feature/timer/src/main/kotlin/com/whakaara/feature/timer/util/TimerUiEvent.kt
@@ -1,0 +1,5 @@
+package com.whakaara.feature.timer.util
+
+sealed class TimerUiEvent {
+    data class ShowSnackbar(val message: String) : TimerUiEvent()
+}
