+++ feature/timer/src/main/kotlin/com/whakaara/feature/timer/util/TimerUiState.kt
@@ -1,0 +1,8 @@
+package com.whakaara.feature.timer.util
+
+data class TimerUiState(
+    val isTimerActive: Boolean = false,
+    val isStart: Boolean = true,
+    val shouldShowPlayButton: Boolean = false,
+    val finishTime: String? = null
+)
